(doc) Update feature description
During a documentation change on docs.chocolatey.org, some suggestions
were made for changes to some newly documented configuration/features
(these had been previously missed). Rather than make the changes to the
docs, which were a copy/paste of the description from the source files,
the decision was taken to first update the documentation in the source
code, and then update the documentation, so that things match up.